Home component modified and Bubble component created
