+++ src/styles/Home.css
@@ -1,0 +1,77 @@
+.zoom-out {
+    /* scale: 1.2; */
+    animation: zoom-out ease forwards;
+    animation-duration: 3s;
+}
+
+.zoom-in {
+    /* scale: 0.7; */
+    animation: zoom-in ease forwards;
+    animation-duration: 3s;
+}
+
+@keyframes zoom-in {
+    0% {
+        transform: scale(0.6, 0.6);
+    }
+
+    100% {
+        transform: scale(1.1, 1.1);
+    }
+}
+
+@keyframes zoom-out {
+    0% {
+        transform: scale(1.1, 1.1);
+    }
+
+    100% {
+        transform: scale(0.6, 0.6);
+    }
+}
+
+.fade-in {
+	opacity: 1;
+	animation-name: fadeInOpacity;
+	animation-iteration-count: 1;
+	animation-timing-function: ease-in;
+	animation-duration: 300ms;
+}
+
+
+.fade-in-scaled {
+    scale: 1.9;
+    /* translate: -40%; */
+	opacity: 1;
+	animation-name: fadeInOpacity;
+	animation-iteration-count: 1;
+	animation-timing-function: ease-in;
+	animation-duration: 300ms;
+}
+
+.fade-out {
+	opacity: 0;
+	animation-name: fadeOutOpacity;
+	animation-iteration-count: 1;
+	animation-timing-function: ease-in;
+	animation-duration: 300ms;
+}
+
+.fade-out-scaled {
+    scale: 1.9;
+    /* translate: -40%; */
+	opacity: 0;
+	animation-name: fadeOutOpacity;
+	animation-iteration-count: 1;
+	animation-timing-function: ease-in;
+	animation-duration: 300ms;
+}
+
+@keyframes fadeOutOpacity {
+	0% {
+		opacity: 1;
+	}
+	100% {
+		opacity: 0;
+	}
+}
